// src/types/subtypes.d.ts
export interface IMDCover {
    w: number;
    h: number;
    b2key: string;
}
// ...
export interface ITrendingManga {
    id: number;
    title: string | null;
    slug: string;
    content_rating: string;
    genres: number[];
    demographic: number | null;
    md_covers: IMDCover[];
}